Adds PROJECT_REVISION to cache key.
This prevents assets from getting cached across projects (when not using
ember-service-worker-project-entangled-registration) and also allows for
a straightforward way to add a "kill switch" mechanism for these caches.

We were already using `VERSION` (which was/is pretty good anyways), but
I intend to add a "kill switch" mechanism based on
`PROJECT_REVISION` (where the consuming project can simply increment a
value and ensure that their caches are fully flushed and whatnot).